fix type mismatch for std::min

+++ actnn/actnn/cpp_extension/quantization_cuda_kernel.cu
@@ -10,7 +10,7 @@
 #include <cuda_runtime.h>
 #include <curand_kernel.h>
 
-#define BLOCK_Y_DIM_MAX ((1l << 16) - 1)
+#define BLOCK_Y_DIM_MAX ((((int64_t)(1)) << 16) - 1)
 
 using torch::IntArrayRef;
 using torch::Tensor;
